fix: PaddingMask the valid_lens option now works fine
