+++ transformerx/layers/masks/__init__.py
@@ -1,3 +1,1 @@
-from .generic import LookAheadMask
-from .core import PaddingMask
-from .core import BaseMask
+from .generic import LookAheadMask, PaddingMask
